refactor Yachtshow component
