+++ src/components/yachts/yachtshow.css
@@ -1,23 +1,23 @@
 span {
   font-weight: bolder;
 }
-
 .name {
   font-size: xx-large;
 }
-
 .single-yacht {
   display: block;
   margin-left: auto;
   margin-right: 5rem;
   width: 60%;
 }
-
 a {
   color: white;
   text-decoration: none;
 }
-
+.text-box,
+.show-button {
+  width: 15rem;
+}
 @media (min-width: 801px) {
   .yacht-card {
     display: flex;
@@ -25,35 +25,28 @@ a {
     align-content: space-around;
     align-items: center;
     justify-content: space-around;
-
     /* margin-top: 1rem; */
   }
-
   .img-wrap {
     margin-top: 10rem;
   }
-
   .card-button {
     margin-right: 1rem;
   }
-
   .show-button {
     color: white;
     background-color: dodgerblue;
     border-color: dodgerblue;
     margin-top: 2rem;
   }
-
   .text-box {
     margin-top: 1rem;
   }
 }
-
 @media (max-width: 800px) {
   .show-button {
     margin-top: 2rem;
   }
-
   .yacht-card {
     display: flex;
     justify-content: center;
